// src/TypeScript/Interfaces/IStaticContent.Dominant.ts
import { IBasicReturn } from './IBasicReturn';

export interface IStaticFirstPage extends IBasicReturn
{
  Id: number;
  Title_PT: string;
  Title_ENG: string;
  Text_PT: string;
  Text_ENG: string;
  Text1_PT: string;
  Text1_ENG: string;
  Text2_PT: string;
  Text2_ENG: string;
  Text3_PT: string;
  Text3_ENG: string;
  Text4_PT: string;
  Text4_ENG: string;
  Text5_PT: string;
  Text5_ENG: string;
  Text6_PT: string;
  Text6_ENG: string;
  Text7_PT: string;
  Text7_ENG: string;
  Text8_PT: string;
  Text8_ENG: string;
  Text9_PT: string;
  Text9_ENG: string;
  Text10_PT: string;
  Text10_ENG: string;
  CreatedBy: string;
  Created: string;
}